Switch to hosted OpenLayers
This is better than keeping a local copy because the hosted
version can always be the latest and greatest. It also reduces
the installation size and solves the problem of where to put
the html and javascript files on each target OS. The only downside
is that it increases the start up time (~7 seconds to get a fully
loaded map, up from ~5 seconds).